+++ src/services/videos/convert/validator.ts
@@ -1,9 +1,11 @@
+import { envConfig } from "src/utils/envConfig";
+
 const verifySignature = (signature: string | null) => {
   if (!signature) {
     return false;
   }
 
-  const webhookSecret = process.env.NOCODB_WEBHOOK_SIGNATURE;
+  const webhookSecret = envConfig.webhookSignature;
 
   return signature == webhookSecret;
 };
