+++ src/utils/envConfig.ts
@@ -6,6 +6,7 @@ const envConfig = {
   cloudinaryName: process.env.CLOUDINARY_NAME,
   cloudinaryApiKey: process.env.CLOUDINARY_API_KEY,
   cloudinaryApiSecret: process.env.CLOUDINARY_API_SECRET,
+  webhookSignature: process.env.WEBHOOK_SIGNATURE,
 };
 
 export { envConfig };
